MM-34419: Fix plugin enable race
A race happens when we try to enable plugins
in an HA environment. There is no loss in functionality
here, but it's a timing bug.

There are two bugs here. One is simple enough, the other
is a bit tricky to solve.

The first one is a case of nested config listeners. `a.InitPlugins`
is called in InitServer, and as well as a config listener. And that
calls `a.SyncPluginsActiveState()`. But inside `a.InitPlugins`,
there is yet another config listener which again calls
`a.SyncPluginsActiveState()`. The first fix is to simply not call
the method again from the nested listener.

The second bug happens because the config changed message is sent
across the cluster only after saving the config locally
in the store.

https://mattermost.atlassian.net/browse/MM-34419

```release-note
NONE
```